Remove fork server helper
The helper was a bad idea, it's too fragile and more confusing than useful.
It also requires to carry one more binary around.
In cases when we can't statically link the target with the fork server,
we can still statically link a little stub with the fork server and
exec the target from it.

PiperOrigin-RevId: 463177806