#Centipede Better way to avoid RLIMIT_AS for sanitized builds. Issue #166
Before this change, we checked for ASAN/MSAN/TSAN at compile-time,
which assume that the runner is built together with the target.
With this change, the checking is done at run-time.

Another similar but independent problem remains.
interceptors in runner_interceptors.cc need to be disabled for ASAN/etc.
This is impossible to do at run time :(

PiperOrigin-RevId: 474690975