Temporarily removed :multi_sanitizer_fuzz_target target while finding a way to make it work
The current implementation of the custom centipede_fuzz_target() rule in build_defs.bzl silently ignores `--config=asan` (or any other *san) in `blaze build --config=asan multi_sanitizer_fuzz_target`, building it without the sanitizer coverage. If the resulting binary is passed to centipede via --extra_binaries, it will not actually do what we expect it to do.

Fixing that is work-in-progress. Until it's complete, remove the target from the BUILD to avoid confusion.

PiperOrigin-RevId: 474951274