Change mutate_batch_size to 2.
This is the only natural size of mutate batch, because effectively
we need to mutate one input and sometimes provide another input
for cross-over. Only if/when we do cross-over with 3+ inputs we will need
larger mutate_batch_size.
This change means that we need to send fewer bytes to Mutate(), which could
be out of process.

However, it may reduce effectiveness on very short runs, since each batch now
works on fewer inputs.
This in fact happens on one of the puzzles, with one of the fixed seeds.
So, bumping the num_runs for puzzles by 2x.

PiperOrigin-RevId: 481270386