test_util.sh: properly catch blaze errors in centipede::maybe_set_var_to_built_executable_path()
maybe_set_var_to_built_executable_path() executes the build command in a subshell, so even when the calling script defines `set -e`, it does not propagate through to the build command, and build errors go unnoticed. The result is that if there is a previously built target binary, it will be erroneously used instead.

PiperOrigin-RevId: 481951802