Ensure that fork server doesn't mess up engine's stdout/stderr
The fork server manipulates the stdout & stderr (e.g. by calling ftruncate() on them), so unless these handles are redirected for the fork server, the parent process's (engine's) stdout & stderr get truncated at random places. This was observed in the output of command_test.cc, where the CommandTest_ForkServer testcase would cut off blaze's output unpredictably from run to run. The previous change in command_test.cc fixed the test. This change fixes the root cause, so we never run into this again.

To achieve that, the command's stdout and/or stderr are now always redirected to /dev/null iff 1) the caller didn't specify an explicit redirection file, and 2) the caller didn't disable the fork server for this command.

PiperOrigin-RevId: 483814026