Retry waiting for fork server if poll() gets interrupted
Unlike `man read.2` (which the old code used to effectively wait for the fork server execution results), `man poll.2` (which the new code uses) is not protected from interrupts and returns errno==EINTR if that happens. So we need to allow at least several `poll` retries to avoid premature failures. For now, this change allows unlimited retries -- we'll see how it goes.

PiperOrigin-RevId: 485456467