Use absl::Time APIs i/o proprietary Timer class to measure execs/s more precisely
Before this change, the fuzzing progress logging would print "exec/s: 0" for elapsed fuzzing times < 1 second. In addition, the code used a proprietary Timer class, added before the absl::Time APIs went open-sourced. The change replaces the former with the latter, and in the process fixes the precision and rounding logic for exec/s.

PiperOrigin-RevId: 494262228