Plumb per-batch timeout to runner
BEFORE: The engine waited for --timeout_per_batch seconds on the runner to post the execution results to the comms pipe, and terminated the process is the runner exceeded that. The runner itself didn't handle the timeout.

AFTER: In addition to the pipe I/O timeout in the engine, the timeout is plumbed to the runner. The runner handles it in the same way it handled the (already plumbed) per-input timeout (writes a failure description to a file and returns EXIT_FAILURE to the engine).

Additionally, the time and RSS limit handling functions have been unified into a single easily extensibe function.

PiperOrigin-RevId: 495744430