Add --timeout_per_batch + Fix incorrect timeout in engine for runner to post results
The `--timeout` flag has a misleading name, and that confusion has propagated through the code base. The engine would wait for `--timeout` seconds for the runner to post execution results to the comms pipe, but inside the runner, `--timeout` is interpreted as "timeout for a single input", not "timeout for the entire batch".

The default `--timeout` is 60 seconds, and the default `--batch_size` is 1000.

In the use cases we've seen so far, 60 seconds was plenty long for batches to finish for a long time during fuzzing, so we've never encountered a problem.

But now, in SiliFuzz ARM experiments with different settings, the average input size grows so much as the corpus matures that the runner eventually exceeds the incorrect timeout expected by the engine, and the engine bails out.

It is tempting to just use `--timeout * --batch_size` as the batch timeout, but `--timeout` should be seen only as a cap on the long tail of the input runtime distribution. If too many inputs in the corpus were to approach it, the corpus would become unusable for practical purposes. Also, with our flag defaults, the default batch timeout would be the impractical 1000 minutes. This change implements a heuristic formula proportional to `--timeout` and `--batch_size` that give ~30 minutes for the default values.

PiperOrigin-RevId: 496723597