Ensure that fork server doesn't mess up engine's stdout/stderr
The fork server manipulates its stdout/stderr, for example by calling ftruncate(), so unless these handles are redirected, the now-shared stdout/stderr of the parent process (the engine) get truncated at random places as well. This was observed in the output of command_test.cc, where the CommandTest_ForkServer testcase would cut off blaze's output unpredictably from run to run.

An earlier change (a few months ago) in command_test.cc fixed that test. This change fixes the root cause, so we never run into this again.

To achieve that, the command's stdout and/or stderr are now always redirected to /dev/null iff 1) the caller didn't specify an explicit redirection file, and 2) the caller didn't disable the fork server for this command by prefixing it with "%f".

PiperOrigin-RevId: 504452186