Bugfix: Runner sometimes failed to write failure description to expected file
The reason was: both WatchdogThread() and RunOneInput() call CheckWatchdogLimits() (the former at regular intervals, the latter once per input). If they happened to do this concurrently, then two instances of WriteFailureDescription() would run concurrently, and attempt to open and write to the same file via 2 different file descriptors. Apparently, this is broken and results in no data being written to the file at all.

PiperOrigin-RevId: 504459422