Workaround for unreliable custom mutator detection that made centipede_main_test flacky
CentipedeDefaultCallbacks::CentipedeDefaultCallbacks() requests a sample mutation of `{{0}}` input to determine whether the target provides an implementation of LLVMFuzzerCustomMutator(). A success is declared when the set of returned mutants is non-empty. However, LLVMFuzzerCustomMutator() implementation in test_fuzz_target.cc sometimes returns an empty set (and should continue to do so, because that's a valid and conformant behavior). The problem was that the passed random `seed` switched between empty and non-empty returns. This change puts additional requirement on `data`, so if it what CentipedeDefaultCallbacks::CentipedeDefaultCallbacks() passes, the returned set is always non-empty, and detection works.
PiperOrigin-RevId: 505760084