Do not log or check in headers.
Some environments (Chromium, at least) have alternative implementations
of macros such as LOG or CHECK which conflict with the absl equivalents
that are used by Centipede.

Remove all such CHECKs, and corresponding header inclusions, from
centipede headers that are intended to be included within fuzzers.

This change may not yet be exhaustive but enables Chromium compilation
of Centipede to get further.