#Centipede change the way CMP features are computed.
Before: for every 'a CMP b' we compute 4 different 6-bit values
based on {a,b}, jam them together into a 24-bit value, mix with
the context (place where CMP happened) and that's out feature.

Now:
* 4 feature domains that represent
  *  a == b
  *  a - b (for small differences)
  *  hamming distance between and b
  *  log2(a-b)
Note: we dropped log2(a^b) because it seems to not provide
any signal in top of hamming and log2(a-b).

This should limit the number of possible CMP features per context
to 1 + 64 + 64 + 64.
Before it was theoretically possible to have much more features per context.

The intent of this change to
a) limit the number of possible CMP features per context.
b) put different signal into different domains so that we can use knobs
to learn their usefulness.

PiperOrigin-RevId: 518098368