If crasher input isn't detected, save out entire batch
If one-by-one execution is unable to reproduce a detected crash, then it's possible multiple inputs are required to trigger the crash.

Rather than trying anything fancy, this CL writes out all of the inputs from a batch that were executed prior to the crash. The intent is to allow manual repro validation.

A TODO is added to experiment with techniques to validate whether executing of inputs can reproduce the crash, such as simply re-executing the whole batch.

PiperOrigin-RevId: 519897703