Distill cospus after fuzzing loop, not before
The distillation was primarily thought of as a dedicated execution mode of the Centipede binary: first run creates a fuzzed corpus, a second run distills that corpus.

However, we've realized that combining the two modes is benefitial, in particular to keep the corpus size in check during continuous fuzzing.

This change gets us half-way there: the distillation logic is moved to after the fuzzing loop. A subsequent change will add shard syncrhonization: the first `--distill_shards` should all wait for the remaining shards to finish fuzzing (i.e. execute their scheduled batches of inputs and exit the fuzzing loop) before proceeding to distillation.

PiperOrigin-RevId: 522176679